Decode datetime instances as UTC/aware PYTHON-125
The C implementation of this is *very* ugly, but might be as good as
we can get. Any improvements/suggestions are appreciated.

This is a BREAKING change if application code was depending on
datetimes being naive (i.e. comparisons and subtractions between aware
and naive datetimes are not supported by Python), but most apps
*probably* aren't. We might need to change this somehow before the
next release...