Direct connection if one host provided PYTHON-176
If only one host is provided try to connect to just
that node and fail otherwise. Previously if the single
host was part of a replica set we would try to find the
set's primary based on other hosts returned by an
'ismaster' query.

We make a direct connection to a single host even if
it's a secondary in a replica set. slaveok is no longer
considered when making a connection.

If multiple hosts are provided we still try to find a
master/primary and fail if we can't.